shipperctl: move `shipperctl admin` commands to their own package
