+++ cmd/shipperctl/cmd/admin/apply.go
@@ -1,4 +1,4 @@
-package cmd
+package admin
 
 import (
 	"fmt"
