+++ cmd/shipperctl/cmd/admin/command.go
@@ -1,6 +1,11 @@
-package cmd
+package admin
 
 import "github.com/spf13/cobra"
+
+var Command = &cobra.Command{
+	Use:   "admin",
+	Short: "administrate Shipper objects",
+}
 
 var clustersCmd = &cobra.Command{
 	Use:   "clusters",
@@ -9,4 +14,5 @@ var clustersCmd = &cobra.Command{
 
 func init() {
 	clustersCmd.AddCommand(applyCmd)
+	Command.AddCommand(clustersCmd)
 }
